refactor(RadioButtonPanel): client/配下をフラット配置に整理する
client/components/ という下位区分に意味がなかったため（hookが存在せず
componentが1つあるだけ）、client/直下にフラットに配置する。

判断基準はCLAUDE.mdの「client 境界が必要なモジュールは client/ に
閉じ込める」節を参照。
