+++ packages/smarthr-ui/src/components/RadioButtonPanel/RadioButtonPanel.tsx
@@ -10,7 +10,7 @@ import { tv } from 'tailwind-variants'
 
 import { RadioButton } from '../RadioButton'
 
-import { ClickablePanel } from './client/components'
+import { ClickablePanel } from './client'
 
 type Props = ComponentProps<typeof RadioButton> & {
   as?: string | ComponentType<any>
